requirements_for_tests.txt
playwright>=1.44.0
pytest-playwright>=0.5.0
pytest-bdd>=7.2.0
pytest-django>=4.8.0
